+++ config/logging.php
@@ -66,6 +66,12 @@ return [
         ],
 
         'userManagement' => [
+            'driver' => 'stack',
+            'channels' => ['daily', 'singleUserManagement', 'slack'],
+            'ignore_exceptions' => false,
+        ],
+
+        'singleUserManagement' => [
             'driver' => 'single',
             'path' => storage_path('logs/userManagement.log'),
             'level' => 'debug',
@@ -85,7 +91,7 @@ return [
             'url' => env('LOG_SLACK_WEBHOOK_URL'),
             'username' => 'Laravel Log',
             'emoji' => ':boom:',
-            'level' => env('LOG_LEVEL', 'critical'),
+            'level' => env('LOG_LEVEL_SLACK', env('LOG_LEVEL', 'critical')),
             'replace_placeholders' => true,
         ],