+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## Unpublished
+### Changed
+- logger channels updated
 
 ## v0.6.6 - 2024-03-15
 - remove button dropdown item blade template, use dedicated css instead